Add unit tests for the schema converter. Move the formatter tests from the guard to a dedicated file for formatters.
